requirements.txt
streamlit>=1.28.0
openai>=1.3.0
anthropic>=0.7.0
pyyaml>=6.0.1
torch
torchaudio
ffmpeg-python>=0.2.0
sqlmodel>=0.0.14
mutagen>=1.47.0
faster-whisper>=1.0.0
torch>=2.0.0
